implement GpuChunk stuff, introduce ~Mem
Implementation of GpuChunkCatalogue. Not thoroughly tested but first tests seem
good.

Introduced new class GpuChunkMem which is handle to memory meant for
GpuChunkHandle but objects will be persistent between uses for different chunks.
So it's something like a memory slot. GpuChunkHandle DOES NOT allocate or free
its arrays itself but rather depends on a GpuChunkMem object for that.

GpuChunkCatalogue has some GpuChunkMem and stores some GpuChunkHandles in them.
It defines the interface for managing the chunks on the GPU.